Module Liquidsoap_lang.Binary_strings_map

val register : string -> unit

Register s as a binary string. The entry is kept alive as long as s is live. Multiple registrations for strings with the same content share a single entry.

val remove : string -> unit

Remove s's content digest from the binary set.

val is_binary : string -> bool

Return true if s's content digest is in the binary set.