Module Liquidsoap_lang.Lang_core

include module type of struct include Value end

type env = (string * Value.t) list

We derive a hash of the environment to invalidate the cache when the builtin env change. We mostly keep name and methods.

and dynamic_methods = Value.dynamic_methods = {

hidden_methods : string list;

methods : string -> Value.t option;

}

val hash_fold_env : Liquidsoap_lang.Term_hash.Ppx_hash_lib.Std.Hash.state -> env -> Liquidsoap_lang.Term_hash.Ppx_hash_lib.Std.Hash.state

val hash_fold_dynamic_methods : dynamic_methods Liquidsoap_lang.Term_hash.Ppx_hash_lib.Std.Hash.folder

val hash_fold_t : Value.t Liquidsoap_lang.Term_hash.Ppx_hash_lib.Std.Hash.folder

val hash_env : env -> Liquidsoap_lang.Term_hash.Ppx_hash_lib.Std.Hash.hash_value

val hash_dynamic_methods : dynamic_methods -> Liquidsoap_lang.Term_hash.Ppx_hash_lib.Std.Hash.hash_value

val hash : Value.t -> Liquidsoap_lang.Term_hash.Ppx_hash_lib.Std.Hash.hash_value

type fun_v = Value.fun_v = {

fun_args : (string * string * Value.t option) list;

fun_env : env;

fun_body : Term.t;

}

type ffi = Value.ffi = {

ffi_args : (string * string * Value.t option) list;

ffi_fn : env -> Value.t;

}

type in_value = [

| `Int of int

| `Float of float

| `String of string

| `Bool of bool

| `Null

| `Custom of Liquidsoap_lang.Term.Custom.t

| `List of Value.t list

| `Tuple of Value.t list

| `Fun of fun_v

| `FFI of ffi

]

val methods : Value.t -> Value.t Runtime_term.Methods.t

val map_methods : Value.t -> (Value.t Runtime_term.Methods.t -> Value.t Runtime_term.Methods.t) -> Value.t

val set_pos : Value.t -> Pos.Option.t -> Value.t

val has_flag : Value.t -> Flags.flag -> bool

val add_flag : Value.t -> Flags.flag -> unit

val remove_flag : Value.t -> Flags.flag -> unit

val is_unit : Value.t -> bool

val fun_id : int Stdlib.Atomic.t

val make : ?pos:Pos.Option.base -> ?methods:Value.t Runtime_term.Methods.t -> ?flags:Flags.flags -> in_value -> Value.t

val string_of_int_value : flags:Flags.flags -> int -> string

val invoke : Value.t -> string -> Value.t

Find a method in a value.

val invokes : Value.t -> string list -> Value.t

Perform a sequence of invokes: invokes x l1;l2;l3;... is x.l1.l2.l3...

val demeth : Value.t -> Value.t

val remeth : Value.t -> Value.t -> Value.t

val split_meths : Value.t -> (string * Value.t) list * Value.t

val compare : Value.t -> Value.t -> int

module type Custom = Value.Custom

module type CustomDef = Value.CustomDef

module MkCustomFromTerm = Value.MkCustomFromTerm

module MkCustom = Value.MkCustom

module RuntimeType = Value.RuntimeType

module Custom = Liquidsoap_lang.Term.Custom

module Methods = Term.Methods

type t = Type.t

type module_name = string

type scheme = Type.scheme

type value = Value.t

Type construction

val int_t : Type.t

val unit_t : Type.t

val float_t : Type.t

val bool_t : Type.t

val string_t : Type.t

val tuple_t : Type.t list -> Type.t

val product_t : Type.t -> Type.t -> Type.t

val record_t : (string * Type.t) list -> Type.t

val optional_record_t : (string * Type.t) list -> Type.t

val method_t : Type.t -> (string * Type.scheme * string) list -> Type.t

val optional_method_t : Type.t -> (string * Type.scheme * string) list -> Type.t

val of_tuple_t : Type.t -> Type.t list

val of_product_t : Type.t -> Type.t * Type.t

val fun_t : Type.t Type_base.argument list -> Type.t -> Type.t

val list_t : Type.t -> Type.t

val of_list_t : Type.t -> Type.t

val nullable_t : Type.t -> Type.t

val univ_t : ?constraints:Type.constr list -> unit -> Type.t

val getter_t : Type.t -> Type.t

val ref_t : Type.t -> Type.t

Value construction

val mk : ?pos:Pos.Option.base -> ?methods:Value.t Liquidsoap_lang.Value.Methods.t -> ?flags:Flags.flags -> Value.in_value -> Value.t

val unit : Value.t

val int : int -> Value.t

val octal_int : int -> Value.t

val hex_int : int -> Value.t

val bool : bool -> Value.t

val float : float -> Value.t

val string : string -> Value.t

val tuple : Value.t list -> Value.t

val product : Value.t -> Value.t -> Value.t

val list : Value.t list -> Value.t

val null : Value.t

val meth : Value.t -> (string * Value.t) list -> Value.t

val record : (string * Value.t) list -> Value.t

val val_fun : (string * string * Value.t option) list -> (env -> Value.t) -> Value.t

val term_fun : (string * string * Value.t option) list -> Term.t -> Value.t

val val_cst_fun : (string * Value.t option) list -> Value.t -> Value.t

val reference : (unit -> Value.t) -> (Value.t -> unit) -> Value.t

Helpers for defining builtin functions.

type proto = (string * t * value option * string option) list

val builtin_type : (string * Type.t * 'a option * 'b) list -> Type.t -> Type.t

val meth_fun : Value.t -> (string * Value.t) list -> Value.t

val mk_module_name : ?base:string -> string -> string

type 'a meth = {

name : string;

scheme : scheme;

descr : string;

value : 'a;

}

val add_builtin : category:Doc.Value.category -> descr:string -> ?flags:Doc.Value.flag list -> ?meth:Value.t meth list -> ?examples:string list -> ?base:string -> string -> (string * Type_base.t * Value.t option * string option) list -> Type.t -> (env -> Value.t) -> string

val add_builtin_value : category:Doc.Value.category -> descr:string -> ?flags:Doc.Value.flag list -> ?base:string -> string -> Value.t -> Type.t -> string

val add_builtin_base : category:Doc.Value.category -> descr:string -> ?flags:Doc.Value.flag list -> ?base:string -> string -> in_value -> Type.t -> string

val add_module : ?base:string -> string -> string

val module_name : 'a -> 'a

val apply_fun : (?pos:Pos.t list -> value -> env -> value) Stdlib.ref

val apply : ?pos:Pos.t list -> value -> env -> value

High-level manipulation of values

val to_unit : Value.t -> unit

val to_bool : Value.t -> bool

val to_bool_getter : Value.t -> unit -> bool

val to_fun : value -> env -> value

val is_fun : Value.t -> bool

val to_string : Value.t -> string

val to_string_getter : Value.t -> unit -> string

val to_float : Value.t -> float

val to_float_getter : Value.t -> unit -> float

val to_int : Value.t -> int

val to_int_getter : Value.t -> unit -> int

val to_num : Value.t -> [> `Float of float | `Int of int ]

val to_list : Value.t -> Value.t list

val to_tuple : Value.t -> Value.t list

val to_option : Value.t -> Value.t option

val to_valued_option : (Value.t -> 'a) -> Value.t -> 'a option

val to_default_option : default:'a -> (Value.t -> 'a) -> Value.t -> 'a

val to_product : Value.t -> Value.t * Value.t

val to_string_list : Value.t -> string list

val to_int_list : Value.t -> int list

val to_getter : Value.t -> unit -> value

val to_ref : Value.t -> (unit -> value) * (Value.t -> unit)

val to_valued_ref : (value -> 'a) -> ('b -> Value.t) -> Value.t -> (unit -> 'a) * ('b -> unit)

val assoc : 'a -> int -> ('a * 'b) list -> 'b

assoc lbl n l returns the nth element in l of which the first component is lbl.

val raise_error : ?bt:Stdlib.Printexc.raw_backtrace -> ?message:string -> pos:Pos.t list -> string -> 'a

val runtime_error_of_exception : bt:Stdlib.Printexc.raw_backtrace -> kind:string -> exn -> Runtime_error.runtime_error

val raise_as_runtime : bt:Stdlib.Printexc.raw_backtrace -> kind:string -> exn -> 'a

val environment : unit -> (string * string) list

module Position : sig ... end

module Stacktrace : sig ... end

val pos_var : string

val pos : (string * Value.t) list -> Pos.t list