Module Liquidsoap_lang.Preprocessor

type tokenizer = unit -> Parser.token * Term_base.parsed_pos

type exp_item =

| String of string

| Expr of tokenizer

| End

exception Found_interpolation

val expand_string : ?fname:string -> (unit -> Parser.token * Term_base.parsed_pos) -> unit -> Parser.token * Term_base.parsed_pos

val int_meth : (unit -> Parser.token * (Stdlib.Lexing.position * Stdlib.Lexing.position)) -> unit -> Parser.token * (Stdlib.Lexing.position * Stdlib.Lexing.position)

Special token in order to avoid 3.s = "a" to be parsed as a float followed by a record.

val dotter : (unit -> Parser.token * 'a) -> unit -> Parser.token * 'a

val uminus : (unit -> Parser.token * 'a) -> unit -> Parser.token * 'a

Change MINUS to UMINUS if the minus is not preceded by a number (or an expression which could produce a number).

val strip_newlines : (unit -> Parser.token * ('a * 'b)) -> unit -> Parser.token * ('a * 'b)

val mk_tokenizer : ?fname:string -> Sedlexing.lexbuf -> unit -> Parser.token * Stdlib.Lexing.position * Stdlib.Lexing.position