Module Liquidsoap_lang.Term_base

Terms and values in the Liquidsoap language.

include module type of struct include Runtime_term end

module Vars = Runtime_term.Vars

Sets of variables.

module Methods = Runtime_term.Methods

type custom = Runtime_term.custom

val hash_fold_custom : Liquidsoap_lang.Term_hash.Ppx_hash_lib.Std.Hash.state -> custom -> Liquidsoap_lang.Term_hash.Ppx_hash_lib.Std.Hash.state

val hash_custom : custom -> Liquidsoap_lang.Term_hash.Ppx_hash_lib.Std.Hash.hash_value

type custom_handler = Runtime_term.custom_handler = {

name : string;

to_string : custom -> string;

to_json : pos:Pos.t list -> custom -> Json.t;

compare : custom -> custom -> int;

typ : Type.t;

}

val hash_fold_custom_handler : Liquidsoap_lang.Term_hash.Ppx_hash_lib.Std.Hash.state -> custom_handler -> Liquidsoap_lang.Term_hash.Ppx_hash_lib.Std.Hash.state

val hash_custom_handler : custom_handler -> Liquidsoap_lang.Term_hash.Ppx_hash_lib.Std.Hash.hash_value

type custom_term = Runtime_term.custom_term = {

value : custom;

handler : custom_handler;

}

val hash_fold_custom_term : Liquidsoap_lang.Term_hash.Ppx_hash_lib.Std.Hash.state -> custom_term -> Liquidsoap_lang.Term_hash.Ppx_hash_lib.Std.Hash.state

val hash_custom_term : custom_term -> Liquidsoap_lang.Term_hash.Ppx_hash_lib.Std.Hash.hash_value

type 'a term = 'a Runtime_term.term = {

t : Type.t;

term : 'a;

flags : Flags.flags;

methods : 'a term Methods.t;

}

val has_flag : 'a term -> Flags.flag -> bool

type ('a, 'b) func_argument = ('a, 'b) Runtime_term.func_argument = {

label : string;

as_variable : string option;

default : 'a option;

typ : 'b;

pos : Pos.t option;

}

val hash_fold_func_argument : 'a 'b. (Liquidsoap_lang.Term_hash.Ppx_hash_lib.Std.Hash.state -> 'a -> Liquidsoap_lang.Term_hash.Ppx_hash_lib.Std.Hash.state) -> (Liquidsoap_lang.Term_hash.Ppx_hash_lib.Std.Hash.state -> 'b -> Liquidsoap_lang.Term_hash.Ppx_hash_lib.Std.Hash.state) -> Liquidsoap_lang.Term_hash.Ppx_hash_lib.Std.Hash.state -> ('a, 'b) func_argument -> Liquidsoap_lang.Term_hash.Ppx_hash_lib.Std.Hash.state

type ('a, 'b) func = ('a, 'b) Runtime_term.func = {

mutable free_vars : Vars.t option;

name : string option;

arguments : ('a, 'b) func_argument list;

body : 'a;

}

type 'a app = 'a * (string * 'a) list

type ('a, 'b) cast = ('a, 'b) Runtime_term.cast = {

cast : 'a;

typ : 'b;

}

val hash_fold_cast : 'a 'b. (Liquidsoap_lang.Term_hash.Ppx_hash_lib.Std.Hash.state -> 'a -> Liquidsoap_lang.Term_hash.Ppx_hash_lib.Std.Hash.state) -> (Liquidsoap_lang.Term_hash.Ppx_hash_lib.Std.Hash.state -> 'b -> Liquidsoap_lang.Term_hash.Ppx_hash_lib.Std.Hash.state) -> Liquidsoap_lang.Term_hash.Ppx_hash_lib.Std.Hash.state -> ('a, 'b) cast -> Liquidsoap_lang.Term_hash.Ppx_hash_lib.Std.Hash.state

type ('a, 'b) common_ast = [

| `Custom of custom_term

| `Tuple of 'a list

| `Null

| `Cast of ('a, 'b) cast

| `Open of 'a * 'a

| `Var of string

| `Seq of 'a * 'a

]

val hash_fold_common_ast : 'a 'b. (Liquidsoap_lang.Term_hash.Ppx_hash_lib.Std.Hash.state -> 'a -> Liquidsoap_lang.Term_hash.Ppx_hash_lib.Std.Hash.state) -> (Liquidsoap_lang.Term_hash.Ppx_hash_lib.Std.Hash.state -> 'b -> Liquidsoap_lang.Term_hash.Ppx_hash_lib.Std.Hash.state) -> Liquidsoap_lang.Term_hash.Ppx_hash_lib.Std.Hash.state -> ('a, 'b) common_ast -> Liquidsoap_lang.Term_hash.Ppx_hash_lib.Std.Hash.state

type 'a invoke = 'a Runtime_term.invoke = {

invoked : 'a;

invoke_default : 'a option;

meth : string;

}

type pattern = [

| `PVar of string list

| `PTuple of string list

]

type 'a let_t = 'a Runtime_term.let_t = {

doc : Doc.Value.t option;

replace : bool;

pat : pattern;

mutable gen : Type.var list;

def : 'a;

body : 'a;

}

type cached_env = Runtime_term.cached_env = {

var_name : int;

var_id : int;

env : Typing.env;

}

type 'a runtime_ast = [

| `Int of int

| `Cache_env of cached_env Stdlib.ref

| `Float of float

| `String of string

| `Bool of bool

| `Let of 'a let_t

| `List of 'a list

| `App of 'a * (string * 'a) list

| `Invoke of 'a invoke

| `Hide of 'a * string list

| `Encoder of string * 'a {encoder_params}1/shadowed/(2cea65def8c82c136df31e5898ba088d)

| `Fun of ('a, Type.t) func

]

type t = ast term

and ast = [

| (t, Type.t) common_ast

| t runtime_ast

]

type encoder = t Runtime_term.encoder

type encoder_params = t Runtime_term.encoder_params

type parsed_pos = Stdlib.Lexing.position * Stdlib.Lexing.position

exception Internal_error of Pos.t list * string

An internal error. Those should not happen in theory...

exception Parse_error of parsed_pos * string

A parsing error.

exception Unsupported_encoder of Pos.t option * string

Unsupported encoder

val conf_debug : bool Stdlib.ref

val conf_debug_errors : bool Stdlib.ref

val debug : bool Stdlib.Lazy.t

Are we in debugging mode?

val profile : bool Stdlib.ref

val ref_t : ?pos:Pos.t -> Type.t -> Type.t

Terms

module Custom = Term_custom

val unit : [> `Tuple of 'a list ]

val is_ground : ([> `Bool of 'b | `Float of 'c | `Int of 'd | `List of 'a list | `Null | `String of 'e | `Tuple of 'a list ] term as 'a) -> bool

val string_of_pat : [< `PTuple of string list | `PVar of string list ] -> string

String representation of terms, (almost) assuming they are in normal form.

val to_string : t -> string

val id : unit -> int

Create a new value.

val free_vars_pat : ([< `PList of 'a list * Vars.elt option * 'a list | `PMeth of 'a option * (Vars.elt * [< `None | `Nullable | `Pattern of 'a ]) list | `PTuple of 'a list | `PVar of Vars.elt list PVar ] as 'a) -> Vars.t

val bound_vars_pat : [< `PTuple of Vars.elt list | `PVar of Vars.elt list ] -> Vars.t

val free_term_vars : ([< `App of 'a term * ('b * 'a term) list | `Bool of 'c | `Cache_env of 'd | `Cast of ('a term, 'e) cast | `Custom of 'f | `Encoder of ('h * [< `Anonymous of 'i | `Encoder of 'g | `Labelled of 'j * 'a term ] list) as 'g | `Float of 'k | `Fun of ('a term, 'l) func | `Hide of 'a term * string list | `Int of 'm | `Invoke of 'a term invoke | `Let of 'a term let_t | `List of 'a term list | `Null | `Open of 'a term * 'a term | `Seq of 'a term * 'a term | `String of 'n | `Tuple of 'a term list | `Var of Vars.elt ] as 'a) term -> Vars.t

val free_fun_vars : ((([< `App of 'b * ('d * 'b) list | `Bool of 'e | `Cache_env of 'f | `Cast of ('b, 'g) cast | `Custom of 'h | `Encoder of ('j * [< `Anonymous of 'k | `Encoder of 'i | `Labelled of 'l * 'b ] list) as 'i | `Float of 'm | `Fun of 'a | `Hide of 'c term * string list | `Int of 'n | `Invoke of 'b invoke | `Let of 'b let_t | `List of 'b list | `Null | `Open of 'b * 'b | `Seq of 'b * 'b | `String of 'o | `Tuple of 'b list | `Var of Vars.elt ] as 'c) term as 'b, 'p) func as 'a) -> Vars.t

val free_vars : ?bound:string list -> (([< `App of 'a * ('c * 'a) list | `Bool of 'd | `Cache_env of 'e | `Cast of ('a, 'f) cast | `Custom of 'g | `Encoder of ('i * [< `Anonymous of 'j | `Encoder of 'h | `Labelled of 'k * 'a ] list) as 'h | `Float of 'l | `Fun of ('a, 'm) func | `Hide of 'b term * string list | `Int of 'n | `Invoke of 'a invoke | `Let of 'a let_t | `List of 'a list | `Null | `Open of 'a * 'a | `Seq of 'a * 'a | `String of 'o | `Tuple of 'a list | `Var of Vars.elt ] as 'b) term as 'a) -> Vars.t

val can_ignore : Type.t -> bool

Values which can be ignored (and will thus not raise a warning if ignored).

Basic checks and errors

exception Unbound of Pos.Option.t * string

Trying to use an unbound variable.

exception Ignored of t

Silently discarding a meaningful value.

exception No_label of t * string * bool * t

No_label (f,lbl,first,x) indicates that the parameter x could not be passed to the function f because the latter has no label lbl. The first information tells whether lbl=x is the first parameter with label lbl in the considered application, which makes the message a bit more helpful.

exception Duplicate_label of Pos.Option.t * string

A function defines multiple arguments with the same label.

exception Missing_arguments of Pos.Option.t * (string * Type.t) list

Some mandatory arguments with given label and typed were not passed to the function during an application.

exception Unused_variable of string * Pos.t

Check that all let-bound variables are used. No check is performed for variable arguments. This cannot be done at parse-time (as for the computation of the free variables of functions) because we need types, as well as the ability to distinguish toplevel and inner let-in terms.

exception Deprecated of string * Pos.t

val check_unused : throw:(bt:Stdlib.Printexc.raw_backtrace -> exn -> unit) -> lib:bool -> (([< `App of 'a * ('c * 'a) list | `Bool of 'd | `Cache_env of 'e | `Cast of ('a, 'f) cast | `Custom of 'g | `Encoder of ('i * [< `Anonymous of 'j | `Encoder of 'h | `Labelled of 'k * 'a ] list) as 'h | `Float of 'l | `Fun of ('a, 'm) func | `Hide of 'b term * string list | `Int of 'n | `Invoke of 'a invoke | `Let of 'a let_t | `List of 'a list | `Null | `Open of 'a * 'a | `Seq of 'a * 'a | `String of 'o | `Tuple of 'a list | `Var of Vars.elt ] as 'b) term as 'a) -> unit

module type Custom = sig ... end

module type CustomDef = sig ... end

module MkCustom (Def : CustomDef) : sig ... end

val make : ?pos:Pos.t -> ?t:Type.t -> ?flags:Flags.flags -> ?methods:ast term Methods.t -> ast -> ast term

Create a new value.

val fresh : handler:Type.Fresh.mapper -> (([< `App of 'a * ('c * 'a) list | `Bool of 'd | `Cache_env of 'e | `Cast of ('a, Type_base.t) cast | `Custom of 'f | `Encoder of ('h * [ `Anonymous of 'i | `Encoder of 'g | `Labelled of 'j * 'a ] list) as 'g | `Float of 'k | `Fun of ('a, Type_base.t) func | `Hide of 'a * 'l | `Int of 'm | `Invoke of 'a invoke | `Let of 'a let_t | `List of 'a list | `Null | `Open of 'a * 'a | `Seq of 'a * 'a | `String of 'n | `Tuple of 'a list | `Var of 'o App Cast Encoder Fun Hide Invoke Let List Null Open Seq Tuple Var ] as 'b) term as 'a) -> 'a