Module Liquidsoap_lang.Term_trim

val trim_type : Type.t -> Type.t

val trim_encoder_params : ([ `Anonymous of 'b | `Encoder of 'c * 'a | `Labelled of 'd * (([< `App of 'e * ('g * 'e) list | `Bool of 'h | `Cache_env of 'i | `Cast of ('e, 'j) Runtime_term.cast | `Custom of 'k | `Encoder of 'c * 'a | `Float of 'l | `Fun of ('e, 'm) Runtime_term.func | `Hide of 'e * string list | `Int of 'n | `Invoke of 'e Runtime_term.invoke | `Let of 'e Runtime_term.let_t | `List of 'e list | `Null | `Open of 'e * 'e | `Seq of 'e * 'e | `String of 'o | `Tuple of 'e list | `Var of 'p App Cast Encoder Fun Hide Invoke Let List Open Seq Tuple ] as 'f) Runtime_term.term as 'e) ] list as 'a) -> 'a

val trim_encoder : (('b * [ `Anonymous of 'c | `Encoder of 'a | `Labelled of 'd * (([< `App of 'e * ('g * 'e) list | `Bool of 'h | `Cache_env of 'i | `Cast of ('e, 'j) Runtime_term.cast | `Custom of 'k | `Encoder of 'a | `Float of 'l | `Fun of ('e, 'm) Runtime_term.func | `Hide of 'e * string list | `Int of 'n | `Invoke of 'e Runtime_term.invoke | `Let of 'e Runtime_term.let_t | `List of 'e list | `Null | `Open of 'e * 'e | `Seq of 'e * 'e | `String of 'o | `Tuple of 'e list | `Var of 'p App Cast Encoder Fun Hide Invoke Let List Open Seq Tuple ] as 'f) Runtime_term.term as 'e) ] list) as 'a) -> 'a

val trim_ast : ([< `App of 'a Runtime_term.term * ('b * 'a Runtime_term.term) list | `Bool of 'c | `Cache_env of 'd | `Cast of ('a Runtime_term.term, 'e) Runtime_term.cast | `Custom of 'f | `Encoder of ('h * [ `Anonymous of 'i | `Encoder of 'g | `Labelled of 'j * 'a Runtime_term.term ] list) as 'g | `Float of 'k | `Fun of ('a Runtime_term.term, 'l) Runtime_term.func | `Hide of 'a Runtime_term.term * string list | `Int of 'm | `Invoke of 'a Runtime_term.term Runtime_term.invoke | `Let of 'a Runtime_term.term Runtime_term.let_t | `List of 'a Runtime_term.term list | `Null | `Open of 'a Runtime_term.term * 'a Runtime_term.term | `Seq of 'a Runtime_term.term * 'a Runtime_term.term | `String of 'n | `Tuple of 'a Runtime_term.term list | `Var of 'o App Cast Encoder Fun Hide Invoke Let List Open Seq Tuple ] as 'a) -> 'a

val trim_term : (([< `App of 'a * ('c * 'a) list | `Bool of 'd | `Cache_env of 'e | `Cast of ('a, 'f) Runtime_term.cast | `Custom of 'g | `Encoder of ('i * [ `Anonymous of 'j | `Encoder of 'h | `Labelled of 'k * 'a ] list) as 'h | `Float of 'l | `Fun of ('a, 'm) Runtime_term.func | `Hide of 'a * string list | `Int of 'n | `Invoke of 'a Runtime_term.invoke | `Let of 'a Runtime_term.let_t | `List of 'a list | `Null | `Open of 'a * 'a | `Seq of 'a * 'a | `String of 'o | `Tuple of 'a list | `Var of 'p App Cast Encoder Fun Hide Invoke Let List Open Seq Tuple ] as 'b) Runtime_term.term as 'a) -> 'a