Module Liquidsoap_lang.Type_base

val debug : bool Stdlib.ref

Show debugging information.

val debug_levels : bool Stdlib.ref

Show variables levels.

val debug_variance : bool Stdlib.ref

Type information comes attached to the AST from the parsing, with appropriate sharing of the type variables. Then the type inference performs in-place unification.

In order to report precise type error messages, we put very dense parsing location information in the type. Every layer of it can have a location. Destructive unification introduces links in such a way that the old location is still accessible.

The level annotation represents the number of abstractions which surround the type in the AST -- function arguments and let-in definitions. It is used to safely generalize types.

Finally, constraints can be attached to existential (unknown, '_a) and universal ('a) type variables.

Types

Type description

type variance = [

| `Covariant

| `Invariant

]

type 'a argument = bool * string * 'a

module R : sig ... end

type custom

type meth_doc = {

meth_descr : string;

category : [ `Method | `Callback ];

}

type t = {

pos : Pos.Option.t;

descr : descr;

}

and constr = {

constr_descr : string;

univ_descr : string option;

satisfied : subtype:(t -> t -> unit) -> satisfies:(t -> unit) -> t -> unit;

}

and constructed = {

constructor : string;

params : (variance * t) list;

}

A type constructor applied to arguments (e.g. source).

and var = {

name : int;

mutable level : int;

mutable constraints : constr Type_constraints.t;

}

Contents of a variable.

and invar =

| Free of var
(*
the variable is free
*)

| Link of variance * t
(*
the variable has bee substituted
*)

and scheme = var list * t

A type scheme (i.e. a type with universally quantified variables).

and meth = {

meth : string;
(*
name of the method
*)

optional : bool;
(*
is the method optional?
*)

scheme : scheme;
(*
type scheme
*)

doc : meth_doc;
(*
documentation
*)

json_name : string option;
(*
name when represented as JSON
*)

}

A method.

and repr_t = {

t : t;

json_repr : [ `Tuple | `Object ];

}

and custom_handler = {

typ : custom;

custom_name : string;

copy_with : (t -> t) -> custom -> custom;

occur_check : (t -> unit) -> custom -> unit;

filter_vars : (var list -> t -> var list) -> var list -> custom -> var list;

repr : (var list -> t -> constr R.t) -> var list -> custom -> constr R.t;

subtype : (t -> t -> unit) -> custom -> custom -> unit;

sup : (t -> t -> t) -> custom -> custom -> custom;

to_string : custom -> string;

}

and var_t = {

id : int;

mutable contents : invar;

}

and descr =

| String

| Int

| Float

| Bool

| Never

| Custom of custom_handler

| Constr of constructed

| Getter of t
(*
a getter: something that is either a t or () -> t
*)

| List of repr_t

| Tuple of t list

| Nullable of t
(*
something that is either t or null
*)

| Meth of meth * t
(*
t with a method added
*)

| Arrow of t argument list * t
(*
a function
*)

| Var of var_t
(*
a type variable
*)

module Constraints : sig ... end

module DS : sig ... end

val string_of_constr : constr -> string

exception NotImplemented

exception Exists of Pos.Option.t * string

exception Unsatisfied_constraint

val unit : descr

module Var : sig ... end

Operations on variables.

module Vars : sig ... end

Sets of variables.

val var_id_atom : int Stdlib.Atomic.t

val var_id : unit -> int

val deref : t -> t

Dereferencing gives you the meaning of a term, going through links created by instantiations. One should (almost) never work on a non-dereferenced type.

val make : ?pos:Pos.Option.base -> descr -> t

Create a type from its value.

val demeth : t -> t

Remove methods. This function also removes links.

val filter_meths : t -> (meth -> bool) -> t

val map_meths : t -> (meth -> meth) -> t

val remeth : t -> t -> t

Put the methods of the first type around the second type.

val invoke : t -> string -> scheme

Type of a method in a type.

val has_meth : t -> string -> bool

Do we have a method with given label?

val invokes : t -> string list -> var list * t

Type of a submethod in a type.

val meth : ?pos:Pos.Option.base -> ?json_name:string -> ?category:[ `Callback | `Method ] -> ?optional:bool -> string -> scheme -> ?doc:string -> t -> t

Add a method to a type.

val meths : ?invoke:(t -> string -> scheme) -> ?pos:Pos.Option.base -> string list -> scheme -> t -> t

Add a submethod to a type.

val split_meths : t -> meth list * t

Split the methods from the type.

val var_name_atom : int Stdlib.Atomic.t

Create a fresh variable.

val var_name : unit -> int

val var : ?constraints:constr list -> ?level:int -> ?pos:Pos.Option.base -> unit -> t

module Fresh : sig ... end

val fresh : t -> t

val to_string_fun : (?generalized:var list -> t -> string) Stdlib.ref

val to_string : ?generalized:var list -> t -> string

String representation of a type.

val string_of_scheme : (var list * t) -> string

val is_fun : t -> bool

val is_source : t -> bool

val custom_types : (string, unit -> t) Stdlib.Hashtbl.t

val register_type : string -> (unit -> t) -> unit

val find_opt_typ : string -> (unit -> t) option

val mk_invariant : t -> unit

val hide_meth : string -> t -> t

val opt_meth : string -> t -> t

val get_meth : string -> t -> meth